refactor: split native lint modules
