+++ src/core/corsa_core/src/lint/mod.rs
@@ -1,0 +1,20 @@
+//! Rust-authored lint rule primitives and built-in type-aware rules.
+
+mod context;
+mod helpers;
+mod registry;
+mod rules;
+#[cfg(test)]
+mod tests;
+mod types;
+
+pub use context::{RuleContext, RustLintRule};
+pub use registry::{LintRuleRegistry, run_default_type_aware_rule};
+pub use rules::{
+    AwaitThenableRule, NoArrayDeleteRule, NoForInArrayRule, NoImpliedEvalRule, NoMixedEnumsRule,
+    NoUnsafeUnaryMinusRule, OnlyThrowErrorRule, PreferFindRule, PreferIncludesRule,
+    PreferRegexpExecRule, UseUnknownInCatchCallbackVariableRule,
+};
+pub use types::{
+    LintDiagnostic, LintFix, LintNode, LintSuggestion, RuleMessage, RuleMeta, TextRange,
+};
